Square Tiles
Added a square tiles map using the Tkinter. It creates a map of squares.
I then added a process that defines the ratio and position of land to
water. If there is land already next to a tile, the probability that it
will be land increases. Right now it travels from top left to bottom
right, giving the the landmass the shape of a peninsule, but the future
plan is to change this so that it works from  out to in.